multiple versions for register layout invaraint extraction

+++ register_info_direct_prompt_chps.py
@@ -3,14 +3,15 @@ from agents.exceptions import InputGuardrailTripwireTriggered
 from agents.extensions.handoff_prompt import RECOMMENDED_PROMPT_PREFIX
 import asyncio  
 import os
+import json
 
 import config
 from defs import UserContext, RegisterList, Manufacturer, RegisterNameList, RegisterInfo
 from agent_tools.tools import get_datasheet_pdf, get_datasheet_section, all_svd_file_paths
 from agent_tools.svd_parsing import get_peripheral_names, get_register_names_for_peripheral
 from agent_tools.pdf_ops import extract_pages_from_pdf
+from agent_tools.find_chapter_with_name import find_chapter_with_name
 
-CURRENT_VS_ID = 'vs_6892501067b08191ac63cc6de06ee629'
 
 def dynamic_instructions(
     context: RunContextWrapper[UserContext], agent: Agent[UserContext]
@@ -64,11 +65,11 @@ info_extraction_agent = Agent[UserContext](
     instructions=dynamic_instructions,
     tools=[
         # get_datasheet_section, 
-        FileSearchTool(
-            max_num_results=2,
-            vector_store_ids=[CURRENT_VS_ID],
-            include_search_results=True,
-        )
+        # FileSearchTool(
+        #     max_num_results=1,
+        #     vector_store_ids=[CURRENT_VS_ID],
+        #     include_search_results=True,
+        # )
     ],
     output_type=RegisterInfo,
 )
@@ -88,7 +89,7 @@ async def main():
 
     global CURRENT_VS_ID
     CURRENT_VS_ID = user_context.vs_id
-    run_number = str(10)
+    run_number = str(4) #str(user_context.run)
     
     output_dir = os.path.join("agent_output", device_name, run_number)
     os.makedirs(output_dir, exist_ok=True)
@@ -97,7 +98,7 @@ async def main():
     # Write the current preprocessing method to a summary file in the output directory
     summary_path = os.path.join(output_dir, "summary")
     with open(summary_path, "w", encoding="utf-8") as summary_file:
-        summary_file.write(f"CURRENT_PREPROCESSING_METHOD: AGENT_W_VECTOR_STORE_2\n")
+        summary_file.write(f"CURRENT_PREPROCESSING_METHOD: {PreprocessingMethod.CHAPTER_SEARCH}\n")
     print(f"Wrote summary to {summary_path}")
 
     # Get all SVD file paths for the device, and find the set of unique peripheral names
@@ -105,33 +106,41 @@ async def main():
     peripheral_names = get_peripheral_names(svd_file_paths)
     print(f"Found {len(peripheral_names)} peripheral names in SVD files")
     
-    peripheral_names = ["RCC"]
     for peripheral_name in peripheral_names:
         user_context.peripheral_name = peripheral_name
         register_names = get_register_names_for_peripheral(svd_file_paths, peripheral_name)
-        print(f"Found {len(register_names)} registers for peripheral {peripheral_name} in SVD files")
-        register_names = ["APB1ENR"]
         for register_name in register_names:
             user_context.register_name = register_name
-            result = await Runner.run(
-                info_extraction_agent,
-                f"""
-                For the register {register_name} in the peripheral {peripheral_name}. Find the
-                    address_offset,
-                    reset_value,
-                    size,
-                    readonly_bits,
-                    write_only_bits,
-                    read_write_bits,
-                    subfields and their enumerated values (if they exist).
-                All the information you provide must be in the datasheet and accurate. If you cannot find a piece of information for a register, leave that field empty.
-                """,
-                context=user_context,
-            )
+            chps_path = os.path.join("devices", device_name, "sections")
+            if os.path.exists(chps_path):
+                try:
+                    chps = find_chapter_with_name(chps_path, peripheral_name)
+                    result = await Runner.run(
+                        info_extraction_agent,
+                        f"""
+                        For the register {register_name} in the peripheral {peripheral_name}. Find the
+                            address_offset,
+                            reset_value,
+                            size,
+                            readonly_bits,
+                            write_only_bits,
+                            read_write_bits,
+                            subfields and their enumerated values (if they exist).
+                        These are relevant pages of the datasheet:
+                        {chps}
+                        All the information you provide must be in the datasheet and accurate. If you cannot find a piece of information for a register, leave that field empty.
+                        """,
+                        context=user_context,
+                    )
 
-            output_path = os.path.join(output_dir, f"{peripheral_name}_{register_name}")
-            with open(output_path, "w", encoding="utf-8") as f:
-                f.write(result.final_output.model_dump_json(indent=2))
+
+                    output_path = os.path.join(output_dir, f"{peripheral_name}_{register_name}")
+                    with open(output_path, "w", encoding="utf-8") as f:
+                        f.write(result.final_output.model_dump_json(indent=2))
+        
+                except FileNotFoundError as e:
+                    print(f"Warning: {e}")
+        
    
    #TODO: Update Run Number, make whole config json based so its easier to update
